Added a space to each occurrence of the value 32000 in the ProbTest0311 dataset to allow for easier parsing in accordance with the following discussion:
From: Roberto Baldacci [r.baldacci@unibo.it]
Sent: 23 March 2012 14:34
To: 'Jerry Swan'
Subject: R: Request for clarification regarding 'Lagrangean Column Generation' paper

Dear Jerry,

first of all note that each entry in the matrices has a length of 5 digit (i.e., “i5” in Fortran). Moreover, some of the entries are equal to 32000. In the attached txt file I have added a space to the fields equal to “32000”. Thus, as the instance has 51 vertex we have 51 x51 x 2=5202 entries in the matrices.

Regards,

Roberto.