final revision - SSBSE Challenge
The SSBSE Challenge is wrapped up, the paper submitted.
For the continuation of the hashCode project, this source tree will be used.
I cleaned it up and marked it up with comments.
I also rewrote the generator to use remarkable constants (centered heptagonal primes) which, experiments suggest, will result in more efficient hashCodes.